chore: remove sqlite dev database
This project has grown beyond it's capabilities and now we have access to the dev database.
