+++ .env.example
@@ -2,8 +2,6 @@ VITE_DISCOVER=
 VITE_FIREBASE_CONFIG=
 VITE_GIS_BASE_URL=
 VITE_PRINT_PROXY_ACCOUNT=
-VITE_DEV_USER_KEY=
-VITE_DEV_USER_TOKEN=
 
-# Set this to your email in the cloud sql db when using start:with-db
+# Optional: set this to your email in the dev database to auto-login as admin
 # DEV_USER_EMAIL=your.email@example.com
