--- functions/knexfile.ts
@@ -1,16 +1,0 @@
-import type { Knex } from 'knex';
-
-const config: { [key: string]: Knex.Config } = {
-  development: {
-    client: 'sqlite3',
-    connection: {
-      filename: './dev.sqlite3',
-    },
-    useNullAsDefault: true,
-    seeds: {
-      directory: './seeds/dev',
-    },
-  },
-};
-
-module.exports = config;
