+++ functions/package.json
@@ -7,18 +7,15 @@
     "build": "npm run clean && tsc",
     "build:watch": "tsc --watch",
     "check": "tsc -b",
+    "check:proxy": "node --import tsx/esm scripts/check-db-proxy.ts",
     "clean": "rimraf lib",
     "deploy": "firebase deploy --only functions",
     "lint": "eslint .",
     "logs": "firebase functions:log",
-    "migrate": "node --import tsx/esm ./node_modules/.bin/knex migrate:up",
-    "migrate:create": "node --import tsx/esm ./node_modules/.bin/knex migrate:make -x ts",
-    "seed": "node --import tsx/esm ./node_modules/.bin/knex seed:run",
-    "seed:create": "node --import tsx/esm ./node_modules/.bin/knex seed:make -x ts",
     "serve": "npm run build && firebase emulators:start --only functions",
     "shell": "npm run build && firebase functions:shell",
     "start": "npm run serve",
-    "start:admin": "node --import tsx/esm scripts/start-with-db-admin.ts"
+    "start:local": "node --import tsx/esm scripts/start-local-dev.ts"
   },
   "dependencies": {
     "@arcgis/core": "^4.34.8",
@@ -39,10 +36,6 @@
     "rimraf": "^6.1.2",
     "tsx": "^4.21.0",
     "typescript": "^5.9.3"
-  },
-  "optionalDependencies": {
-    "better-sqlite3": "^12.6.2",
-    "sqlite3": "^5.1.7"
   },
   "engines": {
     "node": "22"
