+++ package.json
@@ -14,9 +14,9 @@
     "format": "prettier . --write",
     "lint": "eslint . --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-    "start": "concurrently 'cd functions && npm run build:watch' 'npm run dev:firebase-emulators' 'wait-on http://127.0.0.1:5001/ut-dnr-dwr-wri-app-at/us-west3/healthCheck && npm run dev:vite'",
+    "start": "npm run --prefix functions start:local",
     "start:ci": "vite --host",
-    "start:with-db": "npm run --prefix functions start:admin",
+    "start:stack": "concurrently 'cd functions && npm run build:watch' 'npm run dev:firebase-emulators' 'wait-on http://127.0.0.1:5001/ut-dnr-dwr-wri-app-at/us-west3/healthCheck && npm run dev:vite'",
     "storybook": "storybook dev -p 6006",
     "test": "vitest"
   },
